requirements.txt
eq-config==0.4.0
Flask==1.1.1
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
